Move util and exceptions modules out of pycore/
