+++ nixio/entity_with_sources.py
@@ -10,9 +10,9 @@ from __future__ import (absolute_import, division, print_function)
 from .entity_with_metadata import EntityWithMetadata
 from .source import Source
 
-from .pycore import util
+from . import util
 
-from nixio.util.proxy_list import RefProxyList
+from .util.proxy_list import RefProxyList
 
 
 class RefSourceProxyList(RefProxyList):
