+++ nixio/exceptions/__init__.py
