+++ nixio/exceptions/exceptions.py
