+++ nixio/pycore/dimensions.py
@@ -12,7 +12,7 @@ import numpy as np
 
 from ..value import DataType
 from ..dimension_type import DimensionType
-from . import util
+from .. import util
 
 
 class Dimension(object):
