+++ nixio/tag.py
@@ -14,11 +14,11 @@ from .data_array import DataArray
 from .entity_with_sources import EntityWithSources
 from .data_view import DataView
 from .pycore.feature import Feature
-from .pycore.exceptions import (OutOfBounds, IncompatibleDimensions,
-                                UninitializedEntity, InvalidUnit)
+from .exceptions import (OutOfBounds, IncompatibleDimensions,
+                         UninitializedEntity, InvalidUnit)
 from .dimension_type import DimensionType
 from .link_type import LinkType
-from .pycore import util
+from . import util
 
 from nixio.util.proxy_list import ProxyList, RefProxyList
