+++ nixio/util/__init__.py
@@ -1,0 +1,5 @@
+from .util import *
+from . import names
+from . import units
+from . import proxy_list
+from . import find
