+++ nixio/util/names.py
