+++ nixio/util/util.py
